// src/app/models/programation.model.ts
export class ProgramationModel{
    environment: any;
    date: string;
    appointment: string;
    tariff?: number;
    reason?: string;
    doctor?: number;
    doctor_id2?: number;
    id?: number;
    qdetail?: number;
    state?: number;
    patient?: number;
    createdAt?: Date;
    updatedAt?: Date;
}